Support site_key and multiple auth flows
Note that the changes to proxy.php here need to be mirrored to the
copy of proxy.php that we maintain in fr-dev/config in order to
circumvent this version's strict cert checking and allow us to run
civiproxy against a self-signed civi cert.

You should pull down I124ccfb98aba4c279668a2813251d82c29efb70b to
test this with.

This patch is mostly responding to feedback on
https://github.com/systopia/CiviProxy/pull/67

Bug: T349946
Change-Id: I96a85540c1af669d1b7c3c9fbfa7760488dc73ca